+++ guillotina/static/gmi/precache-manifest.5eb5ad1247f3bfcd10958dfe298a840d.js
@@ -1,0 +1,26 @@
+self.__precacheManifest = (self.__precacheManifest || []).concat([
+  {
+    "revision": "da6e91c3ed4428a4d09abcba9b11ceee",
+    "url": "/+manage/index.html"
+  },
+  {
+    "revision": "4d136aeb7284eb7448dc",
+    "url": "/+manage/static/css/main.fcbb1673.chunk.css"
+  },
+  {
+    "revision": "134f0939ac5607f76ef6",
+    "url": "/+manage/static/js/2.1b4e02b5.chunk.js"
+  },
+  {
+    "revision": "d705cb622423d72c5defbf368ca70dcc",
+    "url": "/+manage/static/js/2.1b4e02b5.chunk.js.LICENSE"
+  },
+  {
+    "revision": "4d136aeb7284eb7448dc",
+    "url": "/+manage/static/js/main.7e5e8924.chunk.js"
+  },
+  {
+    "revision": "da4ef6d8a4e37fe5c0fe",
+    "url": "/+manage/static/js/runtime-main.8d3b9f7b.js"
+  }
+]);