+++ CHANGELOG.rst
@@ -7,6 +7,9 @@ CHANGELOG
 - Adding widget field on json schema
   [bloodbare]
 
+- Update orjson to 3.x
+  [waghanza]
+
 
 6.0.19 (2020-12-15)
 -------------------
@@ -16,6 +19,7 @@ CHANGELOG
 
 - Search by text by rank if its on search paramaters otherwise by alpha order
   [bloodbare]
+
 
 6.0.18 (2020-12-05)
 -------------------